renamed geometry submodule to shapes to avoid conflicts
