+++ src/ansys/geometry/core/designer/edge.py
@@ -30,11 +30,11 @@ from pint import Quantity
 
 from ansys.geometry.core.connection import GrpcClient, grpc_curve_to_curve
 from ansys.geometry.core.errors import protect_grpc
-from ansys.geometry.core.geometry.curves.trimmed_curve import ReversedTrimmedCurve, TrimmedCurve
-from ansys.geometry.core.geometry.parameterization import Interval
 from ansys.geometry.core.math.point import Point3D
 from ansys.geometry.core.misc.checks import ensure_design_is_active
 from ansys.geometry.core.misc.measurements import DEFAULT_UNITS
+from ansys.geometry.core.shapes.curves.trimmed_curve import ReversedTrimmedCurve, TrimmedCurve
+from ansys.geometry.core.shapes.parameterization import Interval
 
 if TYPE_CHECKING:  # pragma: no cover
     from ansys.geometry.core.designer.body import Body
