+++ src/ansys/geometry/core/shapes/curves/curve_evaluation.py
