+++ src/ansys/geometry/core/shapes/curves/ellipse.py
@@ -29,14 +29,6 @@ import numpy as np
 from pint import Quantity
 from scipy.integrate import quad
 
-from ansys.geometry.core.geometry.curves.curve import Curve
-from ansys.geometry.core.geometry.curves.curve_evaluation import CurveEvaluation
-from ansys.geometry.core.geometry.parameterization import (
-    Interval,
-    Parameterization,
-    ParamForm,
-    ParamType,
-)
 from ansys.geometry.core.math import (
     UNITVECTOR3D_X,
     UNITVECTOR3D_Z,
@@ -46,6 +38,14 @@ from ansys.geometry.core.math import (
     Vector3D,
 )
 from ansys.geometry.core.misc import Accuracy, Distance
+from ansys.geometry.core.shapes.curves.curve import Curve
+from ansys.geometry.core.shapes.curves.curve_evaluation import CurveEvaluation
+from ansys.geometry.core.shapes.parameterization import (
+    Interval,
+    Parameterization,
+    ParamForm,
+    ParamType,
+)
 from ansys.geometry.core.typing import Real, RealSequence
