+++ src/ansys/geometry/core/shapes/curves/line.py
@@ -28,16 +28,16 @@ from beartype import beartype as check_input_types
 from beartype.typing import Union
 import numpy as np
 
-from ansys.geometry.core.geometry.curves.curve import Curve
-from ansys.geometry.core.geometry.curves.curve_evaluation import CurveEvaluation
-from ansys.geometry.core.geometry.parameterization import (
+from ansys.geometry.core.math import Matrix44, Point3D, UnitVector3D, Vector3D
+from ansys.geometry.core.misc.accuracy import LENGTH_ACCURACY
+from ansys.geometry.core.shapes.curves.curve import Curve
+from ansys.geometry.core.shapes.curves.curve_evaluation import CurveEvaluation
+from ansys.geometry.core.shapes.parameterization import (
     Interval,
     Parameterization,
     ParamForm,
     ParamType,
 )
-from ansys.geometry.core.math import Matrix44, Point3D, UnitVector3D, Vector3D
-from ansys.geometry.core.misc.accuracy import LENGTH_ACCURACY
 from ansys.geometry.core.typing import Real, RealSequence
