+++ src/ansys/geometry/core/shapes/parameterization.py
