+++ src/ansys/geometry/core/shapes/surfaces/surface.py
@@ -24,9 +24,9 @@ from abc import ABC, abstractmethod
 
 from beartype.typing import Tuple
 
-from ansys.geometry.core.geometry.parameterization import Parameterization, ParamUV
-from ansys.geometry.core.geometry.surfaces.surface_evaluation import SurfaceEvaluation
 from ansys.geometry.core.math import Matrix44, Point3D
+from ansys.geometry.core.shapes.parameterization import Parameterization, ParamUV
+from ansys.geometry.core.shapes.surfaces.surface_evaluation import SurfaceEvaluation
 
 
 class Surface(ABC):
