+++ src/ansys/geometry/core/shapes/surfaces/surface_evaluation.py
@@ -23,8 +23,8 @@
 
 from functools import cached_property
 
-from ansys.geometry.core.geometry.parameterization import ParamUV
 from ansys.geometry.core.math import Point3D, UnitVector3D, Vector3D
+from ansys.geometry.core.shapes.parameterization import ParamUV
 from ansys.geometry.core.typing import Real
