+++ tests/test_parameterization.py
@@ -24,8 +24,8 @@ from beartype.roar import BeartypeCallHintParamViolation
 import numpy as np
 import pytest
 
-from ansys.geometry.core.geometry import Interval, Parameterization, ParamForm, ParamType, ParamUV
 from ansys.geometry.core.misc.accuracy import Accuracy
+from ansys.geometry.core.shapes import Interval, Parameterization, ParamForm, ParamType, ParamUV
 
 
 def test_param_uv():
