+++ tests/test_primitives.py
@@ -25,16 +25,6 @@ import numpy as np
 from pint import Quantity
 import pytest
 
-from ansys.geometry.core.geometry import (
-    Circle,
-    Cone,
-    Cylinder,
-    Ellipse,
-    Line,
-    ParamUV,
-    Sphere,
-    Torus,
-)
 from ansys.geometry.core.math import (
     UNITVECTOR3D_X,
     UNITVECTOR3D_Y,
@@ -45,6 +35,7 @@ from ansys.geometry.core.math import (
     Vector3D,
 )
 from ansys.geometry.core.misc import DEFAULT_UNITS, UNITS, Accuracy, Distance
+from ansys.geometry.core.shapes import Circle, Cone, Cylinder, Ellipse, Line, ParamUV, Sphere, Torus
 
 
 def test_cylinder():
